Add files via upload
Using the Top Spotify Listening History Songs in Countries Dataset from Kaggle, I created a Power BI Dashboard to compare & analyse different aspects of the information provided in the dataset. This analysis allowed me to refine the data and create a Power BI dashboard that makes the general public's understanding of the data easier.